chore: move test to proper directory
